Refactor AuthService into interface and implementation; enhance user verification and code validation
