+++ src/main/java/com/proyecto/appclinica/model/dto/auth/VerifyCodeResponse.java
@@ -1,0 +1,6 @@
+package com.proyecto.appclinica.model.dto.auth;
+
+public record VerifyCodeResponse(
+        String message
+) {
+}
