+++ src/main/java/com/proyecto/appclinica/repository/PatientRepository.java
@@ -4,6 +4,9 @@ import com.proyecto.appclinica.model.entity.PatientEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface PatientRepository extends JpaRepository<PatientEntity, Long> {
+    Optional<PatientEntity> findByDni(String identifier);
 }
