+++ src/main/java/com/proyecto/appclinica/service/impl/CredentialsServiceImpl.java
@@ -2,7 +2,9 @@ package com.proyecto.appclinica.service.impl;
 
 import com.proyecto.appclinica.exception.ResourceNotFoundException;
 import com.proyecto.appclinica.model.dto.auth.CredentialsRequestDto;
+import com.proyecto.appclinica.model.entity.ERole;
 import com.proyecto.appclinica.model.entity.PatientEntity;
+import com.proyecto.appclinica.model.entity.RoleEntity;
 import com.proyecto.appclinica.repository.FhirPatientRepository;
 import com.proyecto.appclinica.repository.PatientRepository;
 import com.proyecto.appclinica.service.CredentialsService;
@@ -11,7 +13,9 @@ import org.hl7.fhir.r4.model.ContactPoint;
 import org.hl7.fhir.r4.model.Patient;
 import org.springframework.stereotype.Service;
 
+import java.util.HashSet;
 import java.util.Objects;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -19,9 +23,12 @@ public class CredentialsServiceImpl implements CredentialsService {
 
     private final FhirPatientRepository fhirPatientRepository;
     private final PatientRepository patientRepository;
+    private final RoleServiceImpl roleService;
 
     @Override
     public void createCredentials(CredentialsRequestDto credentialsRequestDto) {
+        Set<RoleEntity> roles = new HashSet<>();
+
         boolean exists = fhirPatientRepository.patientExistsByIdentifier(credentialsRequestDto.getIdentifier());
 
         if (!exists) {
@@ -32,12 +39,15 @@ public class CredentialsServiceImpl implements CredentialsService {
             throw new IllegalArgumentException("Las contraseñas no coinciden");
         }
 
+        RoleEntity role = roleService.findByName(ERole.PATIENT.name());
+        roles.add(role);
+
         Patient patient = fhirPatientRepository.getPatientByIdentifier(credentialsRequestDto.getIdentifier());
-        PatientEntity newPatient = convertPatientToEntity(patient, credentialsRequestDto.getPassword());
+        PatientEntity newPatient = convertPatientToEntity(patient, credentialsRequestDto.getPassword(), roles);
         patientRepository.save(newPatient);
     }
 
-    private PatientEntity convertPatientToEntity(Patient patient, String password) {
+    private PatientEntity convertPatientToEntity(Patient patient, String password, Set<RoleEntity> roles) {
         return PatientEntity.builder()
                 .name(patient.getNameFirstRep().getNameAsSingleString())
                 .lastname(patient.getNameFirstRep().getFamily())
@@ -54,6 +64,7 @@ public class CredentialsServiceImpl implements CredentialsService {
                         .map(ContactPoint::getValue)
                         .orElse(null))
                 .password(password)
+                .roles(roles)
                 .build();
     }
 }
